update admin directory structure
