--- src/app/(auth)/admin/create/[slug]/page.tsx
@@ -1,38 +1,0 @@
-import type { Metadata } from 'next'
-import { createClient } from '@/utils/supabase/server'
-import { redirect } from 'next/navigation'
-import { ProjectForm, ExperienceForm } from '@/components/admin'
-
-export const metadata: Metadata = {
-  title: 'Avery Ondo',
-  description: 'Avery Ondo\'s personal website.',
-}
-
-export default async function PrivatePage({
-  params,
-}: {
-  params: Promise<{ slug: string }>
-}) {
-  const { slug } = await params
-  const supabase = await createClient()
-  const { data, error } = await supabase.auth.getUser()
-  
-  if (error || !data?.user) {
-    redirect('/login/')
-  }
-    
-  return (
-    <main id="main-content">
-
-      <h1 className="text-3xl font-bold">Add a new {slug}</h1>
-
-      {slug === 'project' && (
-        <ProjectForm />
-      )}
-
-      {slug === 'experience' && (
-        <ExperienceForm />
-      )}
-    </main>
-  )
-}
