+++ src/app/(auth)/admin/experience/new/page.tsx
@@ -1,0 +1,22 @@
+import type { Metadata } from 'next'
+import { createClient } from '@/utils/supabase/server'
+import { redirect } from 'next/navigation'
+import { ExperienceForm } from '@/components/admin'
+
+export const metadata: Metadata = {
+  title: 'Avery Ondo - Admin',
+  description: 'Avery Ondo\'s personal website.',
+}
+
+export default async function PrivatePage() {
+
+  const supabase = await createClient()
+  const { data: authData, error: authError } = await supabase.auth.getUser()
+  if (authError || !authData?.user) {
+    redirect('/login/')
+  }
+
+  return <ExperienceForm />;
+
+
+}
