+++ src/components/admin/AdminNav.tsx
@@ -1,5 +1,11 @@
 import { LogoutButton } from '@/components/auth'
-import { PiGearSixDuotone, PiSpeedometerDuotone, PiPuzzlePieceDuotone, PiBuildingOfficeDuotone } from "react-icons/pi";
+import { 
+  PiGearSixDuotone,
+  PiSpeedometerDuotone,
+  PiPuzzlePieceDuotone,
+  PiBuildingOfficeDuotone,
+  PiArticleDuotone 
+} from "react-icons/pi";
 import { createClient } from '@/utils/supabase/server'
 import { NavLink } from '@/components/ui'
 
@@ -13,9 +19,10 @@ export const AdminNav = async () => {
   return (
     <nav className='w-full flex items-center flex-wrap gap-4 justify-between align text-xs'>
       <ul className='flex flex-wrap gap-4 mt-1 justify-center'>
-        <li><NavLink href="/admin/" className='flex items-center gap-1 text-bright-purple dark:text-pink border-0 border-b-4 border-transparent' activeClasses='border-blue'><PiSpeedometerDuotone className="size-6" /> Dashboard</NavLink></li>
-        <li><NavLink href="/admin/list/experience/" className='flex items-center gap-1 text-bright-purple dark:text-pink border-0 border-b-4 border-transparent' activeClasses='border-blue'><PiBuildingOfficeDuotone className="size-6" /> Experience</NavLink></li>
-        <li><NavLink href="/admin/list/projects/" className='flex items-center gap-1 text-bright-purple dark:text-pink border-0 border-b-4 border-transparent' activeClasses='border-blue'><PiPuzzlePieceDuotone className="size-6" /> Projects</NavLink></li>
+        <li><NavLink href="/admin/" className='flex items-center gap-1 text-bright-purple dark:text-pink border-0 border-b-4 border-transparent' activeClasses='border-blue' exactMatchOnly><PiSpeedometerDuotone className="size-6" /> Dashboard</NavLink></li>
+        <li><NavLink href="/admin/pages/" className='flex items-center gap-1 text-bright-purple dark:text-pink border-0 border-b-4 border-transparent' activeClasses='border-blue'><PiArticleDuotone className="size-6" /> Pages</NavLink></li>
+        <li><NavLink href="/admin/experience/" className='flex items-center gap-1 text-bright-purple dark:text-pink border-0 border-b-4 border-transparent' activeClasses='border-blue'><PiBuildingOfficeDuotone className="size-6" /> Experience</NavLink></li>
+        <li><NavLink href="/admin/projects/" className='flex items-center gap-1 text-bright-purple dark:text-pink border-0 border-b-4 border-transparent' activeClasses='border-blue'><PiPuzzlePieceDuotone className="size-6" /> Projects</NavLink></li>
         <li><NavLink href="/admin/profile/" className='flex items-center gap-1 text-bright-purple dark:text-pink border-0 border-b-4 mb-1.5 border-transparent' activeClasses='border-blue'><PiGearSixDuotone className="size-6" /> Settings</NavLink></li>
       </ul>
       <ul className='flex flex-wrap gap-4 mt-1 justify-center'>
